updated to also support PAPI 3 beta release in addition to PAPI 2.
update Sourceme files to use `pwd` as PAPIRUN_HOME and ../papi/src
as the path to the PAPI shared library.

+++ src/hpctoolkit/hpcprof/proffile.hpp
@@ -57,7 +57,7 @@
 //***************************************************************************
 
 // <pc, count>
-typedef std::pair<pprof_off_t, unsigned short> ProfFileEventDatum;
+typedef std::pair<pprof_off_t, unsigned long long> ProfFileEventDatum;
 
 // ProfFileEvent: contains the event name, profiling period and
 // profiling data for the event
